%x and %X fonctionnal

+++ ft_printf.h
@@ -3,11 +3,12 @@
 
 #include <stdarg.h>
 #include <unistd.h>
+#include <stdlib.h>
 
 int		ft_printf(const char *format, ...);
 int		ft_printchar(int c);
 int		ft_printstr (char *str);
-int		ft_printnbr (int n, char *base);
-char	*ft_itoa_base(int n, char *base);
+int		ft_printnbr (long int n, char *base);
+char	*ft_itoa_base(long int n, char *base);
 
 #endif
